requirements.txt
fastapi
uvicorn
requests
python-dotenv
supabase
pydantic
pydantic-settings
google-auth